Фиксы

+++ commands/Owner/say.js
@@ -17,7 +17,9 @@ class Say extends Command {
 	}
 
 	async run(message, args) {
-		if (!args[0]) return message.delete();
+		if (!args[0]) {
+			if (message.deletable) return message.delete();
+		}
 
 		// Arguments split
 		const split = "++";
@@ -27,8 +29,25 @@ class Say extends Command {
 		let attachment = null;
 		if (message.attachments.size > 0) attachment = message.attachments.first();
 
-		if (args[1] && !args[2]) {
-			message.delete();
+		if (!args[1] && !args[2]) {
+			if (message.deletable) message.delete();
+			const saychannel = message.channel;
+			saychannel.sendTyping();
+
+			setTimeout(function () {
+				if (attachment) saychannel.send({
+					content: args[0],
+					files: [{
+						name: attachment.name,
+						attachment: attachment.url
+					}]
+				});
+				else saychannel.send({
+					content: args[0]
+				});
+			}, 2000);
+		} else if (args[1] && !args[2]) {
+			if (message.deletable) message.delete();
 			const saychannel = message.guild.channels.cache.find(channel => channel.name == args[1] || channel.id == args[1]);
 			saychannel.sendTyping();
 
@@ -45,24 +64,8 @@ class Say extends Command {
 				});
 			}, 2000);
 		} else if (args[2]) {
+			if (message.deletable) message.delete();
 			const saychannel = this.client.guilds.cache.find(guild => guild.name == args[2] || guild.id == args[2]).channels.cache.find(channel => channel.name == args[1] || channel.id == args[1]);
-			saychannel.sendTyping();
-
-			setTimeout(function () {
-				if (attachment) saychannel.send({
-					content: args[0],
-					files: [{
-						name: attachment.name,
-						attachment: attachment.url
-					}]
-				});
-				else saychannel.send({
-					content: args[0]
-				});
-			}, 2000);
-		} else {
-			message.delete();
-			const saychannel = message.channel;
 			saychannel.sendTyping();
 
 			setTimeout(function () {
